Added CSVHelper, BhavCopyReader, CSV Files enumerator. Tested Importing csv to Sqlite db

+++ BhavCopyApplication/Model.cs
@@ -9,10 +9,10 @@ namespace BhavCopyApplication
         public int Id { get; set; }
         public string Ticker { get; set; }
         public DateTime date { get; set; }
-        public float Open { get; set; }
-        public float High { get; set; }
-        public float Low { get; set; }
-        public float Close { get; set; }
+        public double Open { get; set; }
+        public double High { get; set; }
+        public double Low { get; set; }
+        public double Close { get; set; }
         public double Volume { get; set; }
 
     }
